@@ -1,21 +1,26 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 23
-    buildToolsVersion "26.0.1"
+    compileSdkVersion 28
+
+    lintOptions {
+        checkReleaseBuilds false
+        // Or, if you prefer, you can continue to check for errors in release builds,
+        // but continue the build even when errors are found:
+        abortOnError false
+    }
 
     dexOptions {
         javaMaxHeapSize "4g"
-
     }
 
     defaultConfig {
         applicationId "org.owasp.seraphimdroid"
-        minSdkVersion 14
-        targetSdkVersion 19
+        minSdkVersion 14 //Junfan should be 14 motify the another app
+        targetSdkVersion 26 //23
         compileOptions {
-            sourceCompatibility JavaVersion.VERSION_1_7
-            targetCompatibility JavaVersion.VERSION_1_7
+            sourceCompatibility JavaVersion.VERSION_1_8
+            targetCompatibility JavaVersion.VERSION_1_8
         }
 
         multiDexEnabled true
@@ -24,23 +29,34 @@ android {
     buildTypes {
         release {
             minifyEnabled false
+//            multiDexKeepFile file('multidex-config.txt')
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.txt'
         }
     }
 }
 
 dependencies {
-    compile fileTree(include: ['*.jar'], dir: 'libs')
+    api fileTree(include: ['*.jar'], dir: 'libs')
 
-    compile 'com.android.support:appcompat-v7:21+'
-    compile 'com.android.support:cardview-v7:21+'
-    compile 'com.android.support:recyclerview-v7:21+'
+    api 'androidx.appcompat:appcompat:1.0.0'
 
-    compile 'com.google.android.gms:play-services:6.1+'
-    compile 'com.google.code.gson:gson:2.3.1'
+    api 'androidx.cardview:cardview:1.0.0'
+    api 'androidx.recyclerview:recyclerview:1.0.0'
+    api 'androidx.legacy:legacy-support-v4:1.0.0'
 
-    compile 'com.android.volley:volley:1.0.0'
-    compile(project(':lucenedroid')) {
+    api 'com.google.android.gms:play-services:6.1+'
+    api 'com.google.code.gson:gson:2.3.1'
+
+    api 'androidx.multidex:multidex:2.0.0'
+
+    api 'com.android.volley:volley:1.0.0'
+    api(project(':lucenedroid')) {
         transitive = false
     }
+
+
+    implementation project(path: ':AnotherMonitor', configuration:'default')
+    implementation project(path: ':networkmonitor')
+
+
 }
