@@ -3,7 +3,7 @@ package org.owasp.seraphimdroid;
 import android.content.Intent;
 import android.net.Uri;
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
+import androidx.fragment.app.Fragment;
 import android.text.Html;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -30,7 +30,7 @@ public class AboutFragment extends Fragment {
 		tvGithub.setTag(GitHubId);
 		tvProjectPage = (TextView) view.findViewById(R.id.tv_project_page_link);
 		tvProjectPage.setTag(ProjectPageId);
-		String devNames = "Nikola Milosevic \n\nFurquan Ahmed \n\nAleksandar Abu-Samra \n\nChetan Karande \n\nKartik Kohli \n\nAditya Dua";
+		String devNames = "Nikola Milosevic \n\nFurquan Ahmed \n\nAleksandar Abu-Samra \n\nChetan Karande \n\nKartik Kohli \n\nAditya Dua \n\nJunfan Huang";
 		tvDevNames.setText(devNames);
 		
 		tvGithub.setText(Html.fromHtml("<u>GitHub</u>"));
