@@ -5,6 +5,7 @@ import android.app.AlarmManager;
 import android.app.AlertDialog;
 import android.app.AppOpsManager;
 import android.app.PendingIntent;
+import android.content.ComponentName;
 import android.content.ContentValues;
 import android.content.Context;
 import android.content.DialogInterface;
@@ -21,15 +22,16 @@ import android.os.Build;
 import android.os.Bundle;
 import android.os.Handler;
 import android.preference.PreferenceManager;
-import android.support.annotation.NonNull;
-import android.support.v4.app.ActionBarDrawerToggle;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentActivity;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentTransaction;
-import android.support.v4.widget.DrawerLayout;
+import androidx.annotation.NonNull;
+import androidx.legacy.app.ActionBarDrawerToggle;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentActivity;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentTransaction;
+import androidx.drawerlayout.widget.DrawerLayout;
 import android.telephony.TelephonyManager;
 import android.util.Log;
+import android.view.LayoutInflater;
 import android.view.MenuItem;
 import android.view.View;
 import android.widget.AdapterView;
@@ -44,6 +46,7 @@ import com.android.volley.VolleyError;
 import com.android.volley.toolbox.JsonObjectRequest;
 import com.android.volley.toolbox.Volley;
 
+import org.anothermonitor.ActivityMain;
 import org.json.JSONException;
 import org.json.JSONObject;
 import org.owasp.seraphimdroid.adapter.DrawerAdapter;
@@ -176,9 +179,10 @@ public class MainActivity extends FragmentActivity {
 		//Set SIM id if not set
 		TelephonyManager telephony = (TelephonyManager) getSystemService(Context.TELEPHONY_SERVICE);
 		String hash = null;
-		if(telephony.getSimSerialNumber()!=null) {
-			hash = telephony.getSimSerialNumber() + telephony.getNetworkOperator() + telephony.getNetworkCountryIso();
-		}
+//		Junfan
+//		if(telephony.getSimSerialNumber()!=null) {
+//			hash = telephony.getSimSerialNumber() + telephony.getNetworkOperator() + telephony.getNetworkCountryIso();
+//		}
 		if(hash!=null && defaults.contains("sim_1")==false) {
 			defaults.edit().putString("sim_1", hash).apply();
 		}
@@ -233,7 +237,7 @@ public class MainActivity extends FragmentActivity {
 
 		adapter = new DrawerAdapter(this, listItems);
 		drawerList.setAdapter(adapter);
-
+        // Junfan
 		drawerList.setOnItemClickListener(new OnItemClickListener() {
 
 			@Override
@@ -242,7 +246,8 @@ public class MainActivity extends FragmentActivity {
 				if (position == 6){
 					getIntent().removeExtra("tags");
 					selectFragment(6);
-				} else {
+				}
+				else{
 					selectFragment(position);
 				}
 			}
@@ -325,7 +330,12 @@ public static void setUnlocked(boolean unlocked) {
 				R.drawable.ic_launcher)));
 		listItems.add(new DrawerItem(itemNames[8], iconList.getResourceId(8,
 				R.drawable.ic_launcher)));
-
+		// Junfan add items
+		listItems.add(new DrawerItem(itemNames[9], iconList.getResourceId(9,
+				R.drawable.ic_launcher)));
+		listItems.add(new DrawerItem(itemNames[10], iconList.getResourceId(10,
+				R.drawable.ic_launcher)));
+		// Junfan
 		iconList.recycle();
 	}
 
@@ -436,7 +446,27 @@ public static void setUnlocked(boolean unlocked) {
 				recordUsage(7);
 				fragment = new org.owasp.seraphimdroid.EducateFragment();
 				break;
-			case 7: {
+			// add cpu usage activity
+			case 7:
+				recordUsage(8);
+				Intent intent_cpu = new Intent(MainActivity.this, ActivityMain.class);
+				startActivity(intent_cpu);
+				drawerList.setItemChecked(position, true);
+				drawerList.setSelection(position);
+				setTitle(itemNames[position]);
+				drawerLayout.closeDrawer(drawerList);
+				fragment = new org.owasp.seraphimdroid.Test();
+				break;
+			case 8:
+				recordUsage(9);
+				Intent intent_network = new Intent(MainActivity.this, ca.rmen.android.networkmonitor.app.main.MainActivity.class);
+				startActivity(intent_network);
+				drawerList.setItemChecked(position, true);
+				drawerList.setSelection(position);
+				setTitle(itemNames[position]);
+				drawerLayout.closeDrawer(drawerList);
+				break;
+			case 9: {
 				if (prevSupportFlag != null) {
 					FragmentManager fragMan = getSupportFragmentManager();
 					fragMan.beginTransaction().remove(prevSupportFlag).commit();
@@ -453,7 +483,7 @@ public static void setUnlocked(boolean unlocked) {
 				prevFrag = frag;
 			}
 			break;
-			case 8:
+			case 10:
 				fragment = new org.owasp.seraphimdroid.AboutFragment();
 				break;
 			default:
