@@ -17,7 +17,7 @@ import android.graphics.Color;
 import android.net.Uri;
 import android.os.Handler;
 import android.os.IBinder;
-import android.support.v4.app.NotificationCompat;
+import androidx.core.app.NotificationCompat;
 import android.util.Log;
 
 import org.owasp.seraphimdroid.LogDetailActivity;
@@ -83,105 +83,115 @@ public class OutGoingSmsRecepter extends Service {
 			super.onChange(selfChange);
 
 			Uri uriSmsUri = Uri.parse(CONTENT_SMS);
-			Cursor cursor = getApplicationContext().getContentResolver().query(
-					uriSmsUri, null, null, null, null);
-			cursor.moveToFirst();
+			try{
+				Cursor cursor = getApplicationContext().getContentResolver().query(
+						uriSmsUri, null, null, null, null);
+				cursor.moveToFirst();
 
-			String type = cursor.getString(cursor.getColumnIndex("type"));
-			String message_id = cursor.getString(cursor.getColumnIndex("_id"));
-			String content = cursor.getString(cursor.getColumnIndex("body"));
+				String type = cursor.getString(cursor.getColumnIndex("type"));
+				String message_id = cursor.getString(cursor.getColumnIndex("_id"));
+				String content = cursor.getString(cursor.getColumnIndex("body"));
 
-			if (type.equals("2") && !message_id.equals(messageId)) {
+				if (type.equals("2") && !message_id.equals(messageId)) {
 
-				messageId = message_id;
+					messageId = message_id;
 
-				// Gettings current running task
-				String topPackageName = "";
-//				if(Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
-//				    UsageStatsManager mUsageStatsManager = (UsageStatsManager) getSystemService("usagestats");
-//				    long time = System.currentTimeMillis();
-//				    // We get usage stats for the last 10 seconds
-//				    List<UsageStats> stats = mUsageStatsManager.queryUsageStats(UsageStatsManager.INTERVAL_DAILY, time - 1000*5, time);
-//				    if(stats != null) {
-//				        SortedMap<Long,UsageStats> mySortedMap = new TreeMap<Long,UsageStats>();
-//				        for (UsageStats usageStats : stats) {
-//				            mySortedMap.put(usageStats.getLastTimeUsed(),usageStats);
-//				        }
-//				        if(mySortedMap != null && !mySortedMap.isEmpty()) {
-//				            topPackageName =  mySortedMap.get(mySortedMap.lastKey()).getPackageName();
-//				        }
-//				    }
-//				}
-//				else {
+					// Gettings current running task
+					String topPackageName = "";
+					//				if(Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
+					//				    UsageStatsManager mUsageStatsManager = (UsageStatsManager) getSystemService("usagestats");
+					//				    long time = System.currentTimeMillis();
+					//				    // We get usage stats for the last 10 seconds
+					//				    List<UsageStats> stats = mUsageStatsManager.queryUsageStats(UsageStatsManager.INTERVAL_DAILY, time - 1000*5, time);
+					//				    if(stats != null) {
+					//				        SortedMap<Long,UsageStats> mySortedMap = new TreeMap<Long,UsageStats>();
+					//				        for (UsageStats usageStats : stats) {
+					//				            mySortedMap.put(usageStats.getLastTimeUsed(),usageStats);
+					//				        }
+					//				        if(mySortedMap != null && !mySortedMap.isEmpty()) {
+					//				            topPackageName =  mySortedMap.get(mySortedMap.lastKey()).getPackageName();
+					//				        }
+					//				    }
+					//				}
+					//				else {
 					ActivityManager am = (ActivityManager) getApplicationContext()
 							.getSystemService(ACTIVITY_SERVICE);
 					topPackageName = am.getRunningAppProcesses().get(0).processName;
-				//}
+					//}
 
 
-				// Check whether the messenger is in foreground or not.
-				if (!SMSAppWhiteList.contains(topPackageName)) {
 
-					Log.d(TAG, topPackageName);
-					PackageManager pm = getPackageManager();
-					ApplicationInfo appInfo = null;
-					String appName = "";
-					boolean was_exception = false;
-					try {
-						appInfo = pm.getApplicationInfo(
-								topPackageName,
-								PackageManager.GET_META_DATA);
-						appName = (String) appInfo.loadLabel(pm);
-					} catch (NameNotFoundException e) {
-						e.printStackTrace();
-						was_exception = true;
+
+					// Check whether the messenger is in foreground or not.
+					if (!SMSAppWhiteList.contains(topPackageName)) {
+
+						Log.d(TAG, topPackageName);
+						PackageManager pm = getPackageManager();
+						ApplicationInfo appInfo = null;
+						String appName = "";
+						boolean was_exception = false;
+						try {
+							appInfo = pm.getApplicationInfo(
+									topPackageName,
+									PackageManager.GET_META_DATA);
+							appName = (String) appInfo.loadLabel(pm);
+						} catch (NameNotFoundException e) {
+							e.printStackTrace();
+							was_exception = true;
+						}
+
+						NotificationCompat.Builder builder = new NotificationCompat.Builder(
+								getApplicationContext());
+						builder.setAutoCancel(true)
+								.setSmallIcon(R.drawable.ic_launcher)
+								.setVibrate(new long[]{300, 500, 300})
+								.setLights(Color.YELLOW, 2000, 3000)
+								.setContentTitle("Suspicious SMS");
+						if (appName.equals(""))
+							appName = "unknown application";
+						builder.setContentText("A suspicious SMS has been sent by "
+								+ appName);
+
+						DatabaseHelper dbHelper = new DatabaseHelper(
+								getApplicationContext());
+						SQLiteDatabase db = dbHelper.getWritableDatabase();
+						ContentValues cv = new ContentValues();
+						cv.put("phone_number",
+								cursor.getString(cursor.getColumnIndex("address")));
+						cv.put("reason", "SMS sent by " + appName
+								+ " without users notices or permission");
+						cv.put("time", CallRecepter.getTime());
+						cv.put("type", LogDetailActivity.SMS_OUT);
+						cv.put("content", content);
+						db.insert(DatabaseHelper.TABLE_SMS_LOGS, null, cv);
+						db.close();
+						dbHelper.close();
+
+						Intent logIntent = new Intent(OutGoingSmsRecepter.this,
+								MainActivity.class);
+						logIntent.putExtra("FRAGMENT_NO", 1);
+						logIntent.putExtra("TAB_NO", 1);
+
+						PendingIntent pLogIntent = PendingIntent.getActivity(
+								getApplicationContext(), 0, logIntent,
+								PendingIntent.FLAG_CANCEL_CURRENT);
+						builder.setContentIntent(pLogIntent);
+
+						// Display Notification
+						((NotificationManager) getSystemService(NOTIFICATION_SERVICE))
+								.notify(0, builder.build());
+
 					}
 
-					NotificationCompat.Builder builder = new NotificationCompat.Builder(
-							getApplicationContext());
-					builder.setAutoCancel(true)
-							.setSmallIcon(R.drawable.ic_launcher)
-							.setVibrate(new long[]{300, 500, 300})
-							.setLights(Color.YELLOW, 2000, 3000)
-							.setContentTitle("Suspicious SMS");
-					if (appName.equals(""))
-						appName = "unknown application";
-					builder.setContentText("A suspicious SMS has been sent by "
-							+ appName);
-
-					DatabaseHelper dbHelper = new DatabaseHelper(
-							getApplicationContext());
-					SQLiteDatabase db = dbHelper.getWritableDatabase();
-					ContentValues cv = new ContentValues();
-					cv.put("phone_number",
-							cursor.getString(cursor.getColumnIndex("address")));
-					cv.put("reason", "SMS sent by " + appName
-							+ " without users notices or permission");
-					cv.put("time", CallRecepter.getTime());
-					cv.put("type", LogDetailActivity.SMS_OUT);
-					cv.put("content", content);
-					db.insert(DatabaseHelper.TABLE_SMS_LOGS, null, cv);
-					db.close();
-					dbHelper.close();
-
-					Intent logIntent = new Intent(OutGoingSmsRecepter.this,
-							MainActivity.class);
-					logIntent.putExtra("FRAGMENT_NO", 1);
-					logIntent.putExtra("TAB_NO", 1);
-
-					PendingIntent pLogIntent = PendingIntent.getActivity(
-							getApplicationContext(), 0, logIntent,
-							PendingIntent.FLAG_CANCEL_CURRENT);
-					builder.setContentIntent(pLogIntent);
-
-					// Display Notification
-					((NotificationManager) getSystemService(NOTIFICATION_SERVICE))
-							.notify(0, builder.build());
 
 				}
+			}
+			// premission denial issue for android 7+
+			catch (Throwable t){
 
 			}
 		}
+
 	}
 
 }
